requirements.txt
python-telegram-bot[job-queue]==20.7
firebase-admin==6.5.0
python-dotenv==1.0.0
requests==2.31.0
colorama==0.4.6
pyfiglet==0.8.post1
tabulate==0.9.0
schedule==1.2.0
cachetools==5.3.2